Finish implementing tests for removing users

+++ app/Actions/Users/RemoveUserFromProjectAction.php
@@ -9,6 +9,7 @@ class RemoveUserFromProjectAction
 {
     public function __invoke(Project $project, User $userToRemove)
     {
+        abort_if($project->owner->id === $userToRemove->id, 400, 'Owner cannot remove self');
         $project->users()->detach($userToRemove);
         return $project;
     }
